web_app/requirements.txt
streamlit
torchvision
mlflow
pillow